Re-build in Dell03 w/ MSVC 16 2019 x64
Move the log and temptemp.txt file output to %LOCALAPPDATA%\sudoku

Few other minor things...

Bumped version.txt to 1.0.1 2020.07.15, and in rc version.

+++ src/Sudo_List.cxx
@@ -337,7 +337,7 @@ VOID Do_WM_NOTIFY( HWND hWnd, WPARAM wParam, LPARAM lParam )
 {
     char buf[256];
     LPNMHDR pnh = (LPNMHDR)lParam;
-    sprintf(buf,"WM_NOTIFY: Code %d - ID %d", pnh->code, pnh->idFrom );
+    sprintf(buf,"WM_NOTIFY: Code %d - ID %d", (int)pnh->code, (int)pnh->idFrom );
     //switch (pnh->code)
 #ifdef ADD_LISTVIEW
     if (g_hListView && g_bDoneInit) {
